pyproject.toml
[project]
name = "lyzr-qdrant-multi-agent-system"
version = "0.1.0"
description = "An enterprise-grade, multi-agent customer support system built with the Lyzr agentic framework and Qdrant vector database."
license = {text = "MIT"}
readme = "README.md"
requires-python = "==3.12.0"
dependencies = [
    "lyzr-automata",
    "qdrant-client",
    "fastembed>=0.2.0",
    "pandas",
    "transformers",
    "torch",
    "accelerate",
    "bitsandbytes",
    "python-dotenv",
    "google-generativeai",
    "ipykernel>=6.30.0",
    "ipywidgets>=8.1.7",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
